requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
requests==2.32.3
PyGithub==2.3.0
GitPython==3.1.43
aiofiles==23.2.1
jinja2==3.1.4
httpx==0.27.0
websockets==12.0
beautifulsoup4==4.12.3
pyjwt==2.8.0
